CHATBOT WITH HISTORY AND CONTEXT/requirements.txt
streamlit
python-dotenv
langchain-core
langchain-groq
langchain-classic